pyproject.toml
[tool.poetry]
name = "tweetyfast"
version = "0.1.0"
description = ""
authors = ["Igor Ruzhilov <iruzhilov@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.10.5"
fastapi = "^0.95.1"
sqlalchemy = {extras = ["all"], version = "^2.0.11"}
pytest = {extras = ["all"], version = "^7.3.1"}
asyncio = "^3.4.3"
asyncpg = "^0.27.0"
uvicorn = {extras = ["all"], version = "^0.22.0"}
environ = "^1.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
env = "^0.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
